Fix compiler warning.
git-svn-id: https://svn.code.sf.net/p/tortoisesvn/code/trunk@22123 154be670-7969-47af-9ab0-c6fb378d2295

+++ src/SubWCRev/SubWCRev.cpp
@@ -388,7 +388,7 @@ int InsertDateW(wchar_t * def, wchar_t * pBuf, size_t & index,
         while (*pEnd != '$')
         {
             pEnd++;
-            if ((pEnd - pBuf)*sizeof(wchar_t) >= (__int64)filelength)
+            if ((__int64)((pEnd - pBuf)*sizeof(wchar_t)) >= (__int64)filelength)
                 return FALSE;   // No terminator - malformed so give up.
         }
         if ((pEnd - pStart) > 1024)
